backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
rembg==2.0.57
Pillow==10.3.0
python-multipart==0.0.9
aiofiles==23.2.1
onnxruntime==1.18.0
numpy==1.26.4
